Move lib files into a govuk_docker folder
In preparation for making it possible to `gem install govuk-docker`.
