+++ lib/govuk_docker/commands/base.rb
