+++ lib/govuk_docker/commands/build.rb
