+++ lib/govuk_docker/commands/compose.rb
