+++ lib/govuk_docker/commands/prune.rb
