+++ lib/govuk_docker/doctor/doctor.rb
