+++ lib/govuk_docker/errors/unknown_service.rb
