+++ lib/govuk_docker/errors/unknown_stack.rb
