+++ lib/govuk_docker/paths.rb
