+++ lib/govuk_docker/setup/docker.rb
