+++ spec/commands/base_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-require_relative "../../lib/commands/base"
+require_relative "../../lib/govuk_docker/commands/base"
 
 describe Commands::Base do
   subject { described_class.new }
