+++ spec/errors/unknown_stack_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-require_relative "../../lib/errors/unknown_stack"
+require_relative "../../lib/govuk_docker/errors/unknown_stack"
 
 describe UnknownStack do
   it "should print a list of available stack in the error message" do
