+++ spec/paths_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-require_relative "../lib/paths"
+require_relative "../lib/govuk_docker/paths"
 
 describe GovukDocker::Paths do
   describe "govuk_root_dir" do
