+++ spec/setup/repo_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-require_relative "../../lib/setup/repo"
+require_relative "../../lib/govuk_docker/setup/repo"
 
 describe Setup::Repo do
   let(:shell_double) { double }
